USB: misc: Add IPC bridge driver
USB IPC bridge driver acts as a transport layer for the IPC router.
The IPC router provides communication between APPS processor and
MODEM processor.

This driver manages an interface that is similar to USB CDC control
interface.  A platform device with "ipc_bridge" name is created when
the MODEM USB is enumerated.  The platform data is populated with
open, read, write and close function pointers.  The suspend/resume
functionality is hidden from IPC router driver.  The driver supports
auto suspend and ensures that USB is not suspended when a write/read
request is in progress.  Platform device is unregistered when MODEM
USB is disconnected.  IPC router driver is expected to synchronize
I/O with disconnect and close the channel from platform device removal
context.

The driver currently supports only one instance of IPC bridge.  The
design is flexible and easily can be extended to support multiple
instances.

CRs-Fixed: 492801
Change-Id: Ida71a0d5d89d946c81eada8e9453cf55b2b06881
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>

Conflicts:
	drivers/usb/misc/Kconfig